Use a subproject to include Cascade in sample
Cascade needs to be included in the 'Embedded Binaries' section of the
sample app in order to get everything to build for the device.

I'm still not sure how to get this to work with workspaces, but if we
include the Cascade project as a sub-project, we can link directly to
the framework without having it end up hardcoded to my harddrive.