Don't partially apply class instance methods
Holy shit did this take a while to figure out. Apparently, you can't
partially apply class instance methods. The worst thing is that if you
do, the compiler doesn't catch it, _and_ it works as long as you're on
the iPhone simulator. Once you run it in the iPad sim, or on the device,
it crashes (or in the case of the iPad sim, hang indefinitely).

Moving this to a global function fixes the issue. But at what cost...