Set the item width to an integer
If this wasn't an integer, but the height was, it led to inaccurate cell
sizes. We were seeing the background bleed through on the right hand
side. If we remove the floor for the height, we would see the background
on the right and the bottom. Flooring both the width and the height
ensures that the cells are sized properly.